# Much faster writeMessage and simpler packed API