--- a/bench/bench.html
+++ b/bench/bench.html
@@ -74,37 +74,32 @@
         function encodeLayers(layers) {
             var pbf = new Pbf();
             for (var i = 0; i < layers.length; i++) {
-                pbf.writeMessage(3, encodeLayer(layers[i]));
+                pbf.writeMessage(3, encodeLayer, layers[i]);
             }
             return pbf.finish();
         }
-        function encodeLayer(layer) {
-            var pbf = new Pbf();
+        function encodeLayer(layer, pbf) {
             pbf.writeStringField(1, layer.name);
 
             for (var i = 0; i < layer.features.length; i++) {
-                pbf.writeMessage(2, encodeFeature(layer.features[i]));
+                pbf.writeMessage(2, encodeFeature, layer.features[i]);
             }
             for (i = 0; i < layer.keys.length; i++) {
                 pbf.writeStringField(3, layer.keys[i]);
             }
             for (i = 0; i < layer.values.length; i++) {
-                pbf.writeMessage(4, encodeValue(layer.values[i]));
+                pbf.writeMessage(4, encodeValue, layer.values[i]);
             }
             pbf.writeVarintField(5, layer.extent);
             pbf.writeVarintField(15, layer.version);
-            return pbf;
         }
-        function encodeFeature(feature) {
-            var pbf = new Pbf();
+        function encodeFeature(feature, pbf) {
             pbf.writeVarintField(1, feature.id);
             pbf.writePackedVarint(2, feature.tags);
             pbf.writeVarintField(3, feature.type);
             pbf.writePackedVarint(4, feature.geometry);
-            return pbf;
         }
-        function encodeValue(value) {
-            var pbf = new Pbf();
+        function encodeValue(value, pbf) {
             if (typeof value === 'string') pbf.writeStringField(1, value);
             else if (typeof value === 'number') {
                 if (value % 1 === 0) { // integer
@@ -112,7 +107,6 @@
                     else pbf.writeSVarintField(6, value);
                 } else pbf.writeDoubleField(3, value);
             } else if (typeof value === 'boolean') pbf.writeBooleanField(7, value);
-            return pbf;
         }
 
     </script>