Closes #19, ref #20.

1. A different `writeMessage` mechanism that accepts function rather than a `Pbf` instance and leads to huge encoding performance improvement.
2. Replaced `readPacked` and `writePacked` with `readPackedVarint`, etc. for better performance (70% faster decoding in browsers) and simpler API.

While the code is a bit more verbose now, the library size grew from 2.99KB to just 3.12KB gzipped — not much. Now on to the benchmarks:

Node encoding:

```
encode vector tile with pbf x 64.65 ops/sec ±1.52% (60 runs sampled) # before
encode vector tile with pbf x 172 ops/sec ±1.16% (82 runs sampled) # after
```

Browser encoding:

```
encode vector tile 10 times: 154ms # before
encode vector tile 10 times: 38ms # after
```

Browser decoding:

```
decode vector tile 10 times: 75ms # before
decode vector tile 10 times: 35ms # after
```